Removed about section code from HomePage and made implemented an AboutSection component
