--- src/Pages/HomePage/HomePage.scss
@@ -1,174 +1,0 @@
-@use '../../styles/partials/animations' as *;
-@use '../../styles/partials/mixins' as *;
-@use '../../styles/partials/variables' as *;
-
-.about {
-  @include section-padding;
-
-  &__title-container {
-    display: flex;
-    gap: 0.75rem;
-    @include tablet {
-      gap: 1rem;
-      flex-direction: column;
-    }
-    @include desktop {
-      gap: 1.25rem;
-    }
-  }
-
-  &__codebox {
-    color: #333;
-    margin: 0 auto;
-    display: flex;
-    flex-direction: column;
-    height: fit-content;
-    background-color: #eee;
-    font-family: monospace;
-    font-size: 0.9rem;
-    padding: 0.75rem 1rem;
-    border-radius: .5rem;
-    border: 1px solid #ccc;
-    box-shadow: $darkshadow1;
-    @include tablet {
-      font-size: 1.1rem;
-      padding: 1.25rem;
-    }
-    @include desktop {
-      padding: 1.75rem;
-      font-size: 1.3rem;
-    }
-
-    &:hover .about__hint {
-      color: #444;
-      transform: scale(1.05)
-    }
-  }
-
-  &__code-label {
-    display: block;
-    width: 100%;
-  }
-
-  &__declaration {
-    color: rgb(1, 81, 173);
-  }
-
-  &__function {
-    color: rgb(207, 17, 17);
-  }
-
-  &__input-wrapper {
-    display: inline-block;
-    color: rgb(53, 110, 66)
-  }
-
-  &__input {
-    cursor: pointer;
-    animation: pulse 1.5s linear infinite;
-    font-size: inherit;
-    color: inherit;
-    width: fit-content;
-    padding: 0;
-    outline: none;
-    background-color: inherit;
-    border: none;
-    font-family: inherit;
-    flex-grow: 0;
-  }
-
-  &__code {
-    display: block;
-    width: 100%;
-    
-    &--single-indent {
-      padding-left: 1rem;
-    }
-
-    &--double-indent {
-      padding-left: 2rem;
-    }
-
-    &--return {
-      color: rgb(1, 81, 173); 
-    }
-  }
-
-  &__myPortfolio-container {
-    padding: 1rem 0;
-  }
-
-  &__button-container {
-    margin-top: 0.75rem;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-
-  &__hint {
-    color: #ccc;
-    padding-right: .5rem;
-    margin: 0;
-    line-height: 100%;
-    transition: transform, color, 0.1s;
-  }
-
-  &__submit {
-    font-family: inherit;
-    font-size: 1.5rem;
-    align-self: flex-end;
-    padding: 0.375rem .75rem;
-    background-color: #fff;
-    border: none;
-    border-bottom: 4px solid #aaa;
-    box-shadow: $darkshadow1;
-    border-radius: .375rem;
-    margin: 0;
-    cursor: pointer;
-    transition: transform 0.2s;
-    &:hover {
-      transform: scale(1.15) translate(-7%, -5%);
-    }
-    @include tablet {
-      padding: 0.375rem 1.75rem;
-      font-size: 1.75rem;
-    }
-    @include desktop {
-      font-size: 2rem;
-    }
-  }
-
-  &__subtitle {
-    @include tablet {
-      font-size: 2rem;
-    }
-    padding: 0 0 1rem;
-    margin: 0;
-  }
-
-  &__copy {
-    margin: 0;
-    font-size: 1.25rem;
-    @include copy-sizing;
-    line-height: 170%;
-  }
-
-  &__face {
-    display: block;
-    text-align: center;
-  }
-
-  &__text-cursor {
-    transform: translateY(0.2rem);
-    content: "";
-    display: inline-block;
-    background-color: #111;
-    margin-left: 0.1rem;
-    width: .1rem;
-    height: 1rem;
-    animation: blink 1s infinite;
-    @include tablet {
-      height: 2rem;
-    }
-  }
-}
